refactor(presentation): update viewmodel and viewmodelfactory
